+++ runtime/include/chpl-comm-callbacks.h
@@ -138,7 +138,7 @@ typedef struct {
       void *raddr;               // Destination address
       size_t size;               // Size of communication
       int32_t commID;            // unique identifier for this get/put
-      int lineno;                // source line of communication
+      int64_t lineno;                // source line of communication
       int32_t filename;          // source file of communication
     } comm;
 
@@ -152,7 +152,7 @@ typedef struct {
       int32_t stridelevels;
       size_t elemSize;
       int32_t commID;           // unique identifier for this get/put
-      int lineno;               // source line of communication
+      int64_t lineno;               // source line of communication
       int32_t filename;         // source file of communication
     } comm_strd;
 
@@ -161,7 +161,7 @@ typedef struct {
       chpl_fn_int_t fid;        //  Function ID
       void *arg;                //  Function arg pointer
       size_t arg_size;          //  Function arg size
-      int lineno;               // source line of communication
+      int64_t lineno;               // source line of communication
       int32_t filename;         // source file of communication
     } executeOn;
