+++ runtime/include/chpl-string.h
@@ -36,7 +36,7 @@ extern chpl_string defaultStringValue;
 
 struct chpl_chpl____wide_chpl_string_s;
 
-chpl_string chpl_wide_string_copy(struct chpl_chpl____wide_chpl_string_s* x, int32_t lineno, int32_t filename);
+chpl_string chpl_wide_string_copy(struct chpl_chpl____wide_chpl_string_s* x, int64_t lineno, int32_t filename);
 
 #define CHPL_SHORT_STRING_SIZE 8
