+++ runtime/include/chpllaunch.h
@@ -79,7 +79,7 @@ void chpl_launcher_no_colocales_error(const char *name);
 int chpl_launch(int argc, char* argv[], int32_t numLocales,
                 int32_t numLocalesPerNode);
 int chpl_launch_handle_arg(int argc, char* argv[], int argNum,
-                           int32_t lineno, int32_t filename);
+                           int64_t lineno, int32_t filename);
 const argDescTuple_t* chpl_launch_get_help(void);
 
 #ifdef __cplusplus
