+++ runtime/include/config.h
@@ -32,7 +32,7 @@ void printHelpMessage(void);
 void initConfigVarTable(void);
 void printConfigVarTable(void);
 void initSetValue(const char* varName, const char* value,
-                  const char* moduleName, int32_t lineno, int32_t filename);
+                  const char* moduleName, int64_t lineno, int32_t filename);
 const char* lookupSetValue(const char* varName, const char* moduleName);
 void installConfigVar(const char* varName, const char* type,
                       const char* moduleName, int isPrivate, int deprecated,
@@ -40,9 +40,9 @@ void installConfigVar(const char* varName, const char* type,
                       const char* unstableMsg);
 
 int handlePossibleConfigVar(int* argc, char* argv[], int argnum,
-                            int32_t lineno, int32_t filename);
+                            int64_t lineno, int32_t filename);
 void parseConfigFile(const char* configFilename,
-                     int32_t lineno, int32_t filename);
+                     int64_t lineno, int32_t filename);
 
 chpl_bool chpl_config_has_value(c_string v, c_string m);
 c_string chpl_config_get_value(c_string v, c_string m);
